Fix "read only" error when playing a song after another finishes
After a song finishes, opening another file to try to play it shows a
"Text is read only" error. This happens because the playback buffer is
in shell mode, and shell-mode sets parts of the buffer to read only.

So when ready-player--playback-buffer clears the playback buffer before
returning it, it needs to use inhibit-read-only.